Refactor: submit input으로 클라이언트 로직 모두 이동
