+++ src/app/admin/(auth)/login/page.tsx
@@ -1,35 +1,10 @@
-'use client';
-
-import { useRouter } from 'next/navigation';
-import { useEffect } from 'react';
-import { useFormState, useFormStatus } from 'react-dom';
-
-import Button from '@/components/button';
 import Input from '@/components/input';
-import { notify } from '@/components/toast';
-import { logInAdmin } from '@/libs/actions/auth';
+import SubmitButton from '@/container/admin/SubmitButton';
 
 const AdminLogInPage = () => {
-  const router = useRouter();
-  const { pending } = useFormStatus();
-  const [state, formAction] = useFormState(logInAdmin, {
-    status: 0,
-    message: '',
-  });
-
-  useEffect(() => {
-    if (state.status === 0) return;
-    if (state.status === 'ok') router.push('/');
-    else {
-      notify('info', state.message);
-    }
-  }, [state, router]);
   return (
     <div className="w-[500px]">
-      <form
-        className="space-y-6"
-        action={formAction}
-      >
+      <form className="space-y-6">
         <div>
           <label
             htmlFor="email"
@@ -63,13 +38,7 @@ const AdminLogInPage = () => {
           </Input>
         </div>
         <div>
-          <Button
-            type="submit"
-            aria-disabled={pending}
-            className="w-full"
-          >
-            관리자 로그인
-          </Button>
+          <SubmitButton authType="login" />
         </div>
       </form>
     </div>
