+++ src/container/admin/SubmitButton/types.ts
@@ -1,0 +1,3 @@
+export interface IProps {
+  authType: 'login' | 'signup';
+}
